Add SQLFluff and Ruff to docs.jsonl
Add Documentation links for SQLFluff and Ruff. Both of these are linters with extensive configuration where having an AI assistant in doing the proper configuration would be a significant improvement for the developer. The difficult part with these is configuration files are typically slowly changing so it is less likely that developers know the patterns and configurations by heart. They are more likely to spend significant time the documentation to get the configurations correct.